gateway/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
redis[hiredis]==5.0.4
asyncpg==0.29.0
websockets==12.0
python-dotenv==1.0.1
python-ulid==2.2.0      # nicer IDs than raw UUID
SQLAlchemy==2.0.30 
aiohttp>=3.12.7